test: fold records helpers back into the integration test
Drop the shared helper module and the unit tests that exercised it. The
stabilization fixes now live inline in the integration test.
